backend/requirements.txt
fastapi
uvicorn
requests
openai
pypdf
python-multipart
bs4
faiss-cpu
weasyprint
jinja2
boto3
pyjwt
pycognito
python-dotenv
psycopg2-binary